CustomDM: Fix for some level 1 scrolls improperly stating that the player is not skilled enough to learn the spell it contains.

+++ Source/ACE.Server/WorldObjects/Player_Skills.cs
@@ -517,11 +517,11 @@ namespace ACE.Server.WorldObjects
             return -1;
         }
 
-        private const uint magicSkillCheckMargin = 50;
+        private const int magicSkillCheckMargin = 50;
 
         public bool CanReadScroll(Scroll scroll)
         {
-            var power = scroll.Spell.Power;
+            var power = (int)scroll.Spell.Power;
 
             // level 1/7/8 scrolls can be learned by anyone?
             if (Common.ConfigManager.Config.Server.WorldRuleset != Common.Ruleset.CustomDM && (power < 50 || power >= 300)) return true;
